requirements.txt
defusedxml==0.5.0
diff-match-patch==20181111
Django==1.11.18
django-import-export==1.2.0
et-xmlfile==1.0.1
jdcal==1.4
odfpy==1.4.0
openpyxl==2.5.14
Pillow==5.4.1
pyBarcode==0.7
pytz==2018.9
PyYAML==3.13
reportlab==3.5.13
tablib==0.12.1
unicodecsv==0.14.1
xlrd==1.2.0
xlwt==1.3.0
